I tried adding a scroll near to Archive.
Tried adding a scroll next to Archive. To do this, I needed to get a reference to the field if it is overflow. The problem was that the Material UI uses lazy loading and because of that the block needs to appear and then you can get its reference. And the problem is that if you display the reference object, it may show null, but there is something inside there.